[Factory] We can now build libSystem using our new build system

+++ Factory/CrossTools/BuildlibSystem.sh
@@ -1,17 +1,5 @@
 #!/usr/bin/env bash
 
-
-# === CONFIGURATION AND SETUP ===
-
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-
-echo $DIR
-
-CROSSTOOLS="$DIR/Cross"
-ARCH="x86_64"
-TOOLS=$(realpath "$DIR/../Toolchain")/Tools
-UTOPIA_TARGET="$ARCH-trip-linux-utopia"
-
 SHA256SUM="sha256sum"
 REALPATH="realpath"
 MAKE="make"
@@ -66,42 +54,51 @@ fi
 
 # In reality we just cp the files from Core/Libraries/libSystem
 
-ROOT=$(${REALPATH} $DIR/../../)
-cp -vr $ROOT/Core/Libraries/libSystem $DIR/Tarballs/libSystem
+cp -vr $ROOT/Core/Libraries/libSystem $SOURCES_DIR/libSystem
 
 # == BUILD ==
 
-mkdir -p $DIR/Build/musl
+mkdir -p $BUILD_DIR/libSystem
 
-pushd "$DIR/Build/musl"
-    export PATH=$DIR/Cross/bin:/bin:/usr/bin
-    buildstep musl/configure $DIR/Tarballs/libSystem/configure CROSS_COMPILEs=${UTOPIA_TARGET}- \
-    --prefix=/ \
-    --target=${UTOPIA_TARGET}
+pushd "$BUILD_DIR/libSystem"
+    if [ ! -f ".built_$PHASE" ]; then
+        $SOURCES_DIR/libSystem/configure CROSS_COMPILEs=${TARGET_TRIPLE}- \
+        --prefix=/ \
+        --target=${TARGET_TRIPLE}
 
-    make
-    DESTDIR=$CROSSTOOLS make install
+        make
+        DESTDIR=$CROSS_LOCATION make install
+        touch .built_$PHASE
+    else
+        echo "libSystem already built"
+    fi
 popd
 
 # == FIX BROKEN LINKS ==
 
-pushd "$CROSSTOOLS"
-    mkdir -v usr
-    ln -sfv ../include ./usr/include
+pushd "$CROSS_LOCATION"
+    if [ ! -f ".cleaned_$PHASE" ]; then
+        mkdir -pv usr
+        ln -sfv ../include ./usr/include
 
-    rm -vf lib/ld-musl-${ARCH}.so.1
-    ln -sfv libSystem.so lib/ld-musl-${ARCH}.so.1
-    # For legacy applications and portability
-    ln -sfv libSystem.so lib/libc.so
+        rm -vf lib/ld-musl-${CONFIG_ARCHITECTURE}.so.1
+        ln -sfv libSystem.so lib/ld-musl-${CONFIG_ARCHITECTURE}.so.1
+        # For legacy applications and portability
+        ln -sfv libSystem.so lib/libc.so
 
-    ln -sv ../lib/ld-musl-$ARCH.so.1 bin/ldd
+        ln -sfv ../lib/ld-musl-$CONFIG_ARCHITECTURE.so.1 bin/ldd
 
-    mkdir -pv etc
-    if [ -f $CROSSTOOLS/etc/ld-musl-$ARCH.path ]; then
-        rm -rf $CROSSTOOLS/etc/ld-musl-$ARCH.path
+        mkdir -pv etc
+        if [ -f $CROSS_LOCATION/etc/ld-musl-$CONFIG_ARCHITECTURE.path ]; then
+            rm -rf $CROSS_LOCATION/etc/ld-musl-$CONFIG_ARCHITECTURE.path
+        fi
+        # cat didnt work very well here...
+        touch $CROSS_LOCATION/etc/ld-musl-$CONFIG_ARCHITECTURE.path
+        echo "$CROSS_LOCATION/lib" >> $CROSS_LOCATION/etc/ld-musl-$CONFIG_ARCHITECTURE.path
+        echo "$TOOLS_LOCATION/lib" >> $CROSS_LOCATION/etc/ld-musl-$CONFIG_ARCHITECTURE.path
+
+        touch .cleaned_$PHASE
+    else
+        echo "libSystem broken links have already been fixed"
     fi
-    # cat didnt work very well here...
-    touch $CROSSTOOLS/etc/ld-musl-$ARCH.path
-    echo "$CROSSTOOLS/lib" >> $CROSSTOOLS/etc/ld-musl-$ARCH.path
-    echo "$TOOLS/lib" >> $CROSSTOOLS/etc/ld-musl-$ARCH.path
 popd
